+++ packages/ui/Marquee/MarqueeDecorator.js
@@ -147,12 +147,12 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 	const forwardEnter = forward(enter);
 	const forwardLeave = forward(leave);
 
-	const determineTextDirection = (node, rtl, forceDirection, noRtlOverride) => {
+	const determineTextDirection = (node, rtl, forceDirection) => {
 		// Text directionality is a function of locale direction (rtl), content (node.textContent),
 		// and props (forceDirection) in increasing order of significance.
 		if (forceDirection) {
-			rtl = forceDirection === 'rtl';
-		} else if (!noRtlOverride && node) {
+			rtl = forceDirection === 'locale' ? rtl : forceDirection === 'rtl';
+		} else if (node) {
 			rtl = marqueeDirection(node.textContent) === 'rtl';
 		}
 
@@ -194,14 +194,17 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 			/**
 			 * Forces the `direction` of the marquee.
 			 *
-			 * Valid values are `'rtl'` and `'ltr'`. This includes non-text elements as well.
+			 * Valid values are `'rtl'`, `'ltr'`, and `'locale'`. This includes non-text elements as well.
 			 * The default behavior, if this prop is unset, is to evaluate the text content for
 			 * directionality using {@link i18n/util.isRtlText}.
+			 *
+			 * If `'locale'`, the `direction` is determined by the locale, same as {@link ui/Marquee.MarqueeDecorator.rtl}
+			 * In other words, it will not consider the text content for determining the direction.
 			 *
 			 * @type {String}
 			 * @public
 			 */
-			forceDirection: PropTypes.oneOf(['rtl', 'ltr']),
+			forceDirection: PropTypes.oneOf(['rtl', 'ltr', 'locale']),
 
 			/**
 			 * The current locale as a
@@ -286,16 +289,6 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 			marqueeSpeed: PropTypes.number,
 
 			/**
-			 * If true, the rtl value is not overridden by the result of the
-			 * marqueeDirection function.
-			 * In other words, the rtl value is not determined by the content.
-			 *
-			 * @type {Boolean}
-			 * @public
-			 */
-			noRtlOverride: PropTypes.bool,
-
-			/**
 			 * Used to signal for a remeasurement inside of marquee.
 			 *
 			 * The value must change for the remeasurement to take place. The value
@@ -331,7 +324,7 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 				animating: false,
 				overflow: 'ellipsis',
 				promoted: false,
-				rtl: determineTextDirection(null, props.rtl, props.forceDirection, props.noRtlOverride)
+				rtl: determineTextDirection(null, props.rtl, props.forceDirection)
 			};
 			this.sync = false;
 			this.timerState = TimerState.CLEAR;
@@ -357,7 +350,7 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 		}
 
 		componentDidUpdate (prevProps) {
-			const {children, disabled, forceDirection, locale, marqueeOn, marqueeDisabled, marqueeSpacing, marqueeSpeed, noRtlOverride, rtl} = this.props;
+			const {children, disabled, forceDirection, locale, marqueeOn, marqueeDisabled, marqueeSpacing, marqueeSpeed, rtl} = this.props;
 
 			let forceRestartMarquee = false;
 
@@ -382,7 +375,6 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 				prevProps.marqueeOn !== marqueeOn ||
 				prevProps.marqueeDisabled !== marqueeDisabled ||
 				prevProps.marqueeSpeed !== marqueeSpeed ||
-				prevProps.noRtlOverride !== noRtlOverride ||
 				prevProps.forceDirection !== forceDirection
 			) {
 				this.cancelAnimation();
@@ -831,7 +823,7 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 
 		validateTextDirection () {
 			this.setState((state, props) => {
-				const rtl = determineTextDirection(this.node, props.rtl, props.forceDirection, props.noRtlOverride);
+				const rtl = determineTextDirection(this.node, props.rtl, props.forceDirection);
 				return state.rtl === rtl ? null : {rtl};
 			});
 		}
@@ -873,7 +865,6 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 			delete rest.marqueeSpacing;
 			delete rest.marqueeResetDelay;
 			delete rest.marqueeSpeed;
-			delete rest.noRtlOverride;
 			delete rest.remeasure;
 			delete rest.rtl;
 
@@ -912,7 +903,6 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 			delete props.marqueeSpacing;
 			delete props.marqueeResetDelay;
 			delete props.marqueeSpeed;
-			delete props.noRtlOverride;
 			delete props.remeasure;
 			delete props.rtl;